// AllPay-FrontEnd/src/domain/models/acuerdos_pago/cuotas/obtenerCuotasPlanPago.model.ts
export interface DetallePlanPago {
    nro_solicitud: number;
    tipo_documento_recaudador: string;
    numero_documento_recaudador: string;
    nombre_recaudador: string;
    fecha_solicitud: string;
    estado: string;
    id_plan_pago: number;
    numero_plan_pago: number;
    numero_cuota: number;
    fecha_pago: string | null;
    Nro_factura: number;
    cuota_fomento: number;
    valor_factura: number;
    id_cuota_acuerdo_pago: number;
    doc_pago_liquidacion?: string | null;
}

export interface ObtenerDetallesPlanPagoResponse {
    success: boolean;
    detail: string;
    data: DetallePlanPago[];
} 